This revision makes the following changes:      - The 'none' basic geometry type has been added.      - The center() method has been added to geometries.      - The centerUnitOuterNormal() has been added to intersections.      - The GridInterfaceEuler has been changed to use these methods.      - The gie_test program now also tests SGrid.
The changes that have been made to the Dune core have also been sent
for evaluation to Christian Engwer, with the intention that they will
eventually be merged into dune-common and dune-grid.

After this update, you should update the core modules (maybe even from
scratch) and run the patch script again.

The versions of Dune core modules that work with this revision are:
dune-common 5763
dune-grid 6059
dune-istl 1143



git-svn-id: http://public.ict.sintef.no/openrs/svn/trunk@708 cebe8a93-341e-4660-8632-17bdd9c9fbcb